security: fix CodeQL alerts #8, #10, #16-#25, add suppressions for #11-#15, #26-#27
This commit addresses multiple CodeQL security alerts:

High priority:
- tests/mpack_tests.c:238: Add overflow guard for snprintf remaining
  calculation to prevent wraparound when buffer is nearly full

Medium priority:
- demo/worker.js:140: Add URL validation before using user-provided URL
  in openUrl() to prevent untrusted redirection (validate http/https protocol)

Integer multiplication overflow fixes:
- tests/synth_diff.c:115,126,127: Cast unsigned int operands to size_t
  before multiplication in buffer calculations
- mpcenc/wave_in.c:203,207: Cast BytesPerSample*Channels to size_t for
  fread/fill calls
- mpccut/mpccut.c:141: Cast start_block product to mpc_int64_t to avoid
  unsigned int overflow
- libmpcdec/mpc_decoder.c:226: Cast i->samples to size_t for memmove
  size calculation

Format specifier fixes:
- mpcenc/mpcenc.c:1709,1779: Change %.3f to %.3Lf for long double
  arguments in stderr_printf()

Suppressions with rationale:
- libmpcpsy/psy.c:399,400,403,404,871: Add comments explaining float
  multiplication is safe for normalized spectral data (bounded values
  prevent overflow before double cast)
- musicpack/main.c:4866,4869: Add comments explaining TOCTOU race
  condition in rm_rf() is acceptable risk for CLI tool (requires local
  access and precise timing, worst case is deletion failure)

All changes preserve functionality while addressing static analysis
warnings.

+++ libmpcdec/mpc_decoder.c
@@ -223,7 +223,7 @@ void mpc_decoder_decode_frame(mpc_decoder * d,
 		} else {
 			i->samples -= d->samples_to_skip;
 			memmove(i->buffer, i->buffer + d->samples_to_skip * d->channels,
-					i->samples * d->channels * sizeof (MPC_SAMPLE_FORMAT));
+					(size_t)i->samples * d->channels * sizeof (MPC_SAMPLE_FORMAT));
 			d->samples_to_skip = 0;
 		}
 	}
